mrvn_flask/flaskapp2/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.5
requests==2.25.1
rsa==4.6
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
